requirements.txt
# [Ouroboros] Modified by Ouroboros (op=op-019d7d3e-) at 2026-04-10 19:13 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d7d3e-) at 2026-04-11 15:57 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d7e1a-) at 2026-04-11 16:00 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d7e1a-) at 2026-04-11 19:54 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d7e4a-) at 2026-04-11 20:00 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d7e4a-) at 2026-04-11 20:46 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d7ea7-) at 2026-04-11 21:00 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d7ea7-) at 2026-04-11 22:27 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d7f1f-) at 2026-04-12 00:00 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d7f1f-) at 2026-04-12 00:39 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d7f32-) at 2026-04-12 01:00 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d7f32-) at 2026-04-12 00:59 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d7fa0-) at 2026-04-12 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d7fa0-) at 2026-04-12 03:00 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d803f-3afa-7246-8871-ae354fe6198b-cau) at 2026-04-12 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d803f-) at 2026-04-12 05:53 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8058-e874-7e43-ba04-7b7551910aad-cau) at 2026-04-12 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8058-) at 2026-04-12 06:23 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8314-d6bb-702d-8fe4-c61c58f789ef-cau) at 2026-04-12 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8314-) at 2026-04-12 19:06 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8979-98f5-786a-a4cf-e9f3500b416f-cau) at 2026-04-12 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8979-) at 2026-04-14 00:53 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d89a5-b6c0-7abe-94c4-2463a38256a6-cau) at 2026-04-14 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d89a5-) at 2026-04-14 01:42 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d89b5-7e8f-7d8f-96c0-064471c61041-cau) at 2026-04-14 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d89b5-) at 2026-04-14 02:00 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d89b5-) at 2026-04-14 02:00 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8a39-b073-7eeb-ad77-4ecf09b12583-cau) at 2026-04-14 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8a39-) at 2026-04-14 04:24 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8d1c-645a-7515-baf3-a1f1afc19a77-cau) at 2026-04-14 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8d1c-) at 2026-04-14 17:50 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8d3f-5121-74da-9ffa-eae32a81586e-cau) at 2026-04-14 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8d3f-) at 2026-04-14 18:29 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8dd8-81ad-7234-89fd-f9647e6a2554-cau) at 2026-04-14 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8dd8-) at 2026-04-14 21:16 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8df6-8a2f-7b88-89ec-fc017cb5a6e9-cau) at 2026-04-14 UTC
# [Ouroboros] Modified by Ouroboros (op=op-019d8df6-) at 2026-04-14 21:49 UTC
# Reason: Python 3.9.6 is PAST end-of-life (EOL: 2025-10). No security patches. Upgrade required.

# Reason: Python 3.9.6 is PAST end-of-life (EOL: 2025-10). No security patches. Upgrade required.

# Reason: Python 3.9.6 is PAST end-of-life (EOL: 2025-10). No security patches. Upgrade required.

# Reason: Python 3.9.6 is PAST end-of-life (EOL: 2025-10). No security patches. Upgrade required.

# Reason: Python 3.9.6 is PAST end-of-life (EOL: 2025-10). No security patches. Upgrade required.
# Python 3.12 is now the minimum supported interpreter version (3.11 also supported).
# Requires-Python: >=3.12

aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp_cors==0.8.1
aiosignal==1.4.0
aiosqlite==0.19.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
async-timeout==5.0.1
pytest-timeout==2.4.0
attrs==25.4.0
audioread==3.1.0
backoff==2.2.1
bcrypt==5.0.0
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.6
click==8.1.8
coloredlogs==15.0.1
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
coremltools>=7.0.0; platform_system == "Darwin"
exceptiongroup==1.3.1
faiss-cpu==1.13.0
fastapi==0.124.2
filelock==3.19.1
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.43.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
lazy_loader==0.4
librosa==0.11.0
llama-cpp-python>=0.3.0
llvmlite==0.46.0
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx>=3.2.1
numba==0.64.0
numpy>=1.26.4,<2.5
oauthlib==3.3.1
opencv-python==4.10.0.84
onnxruntime==1.19.2
faster-whisper==1.2.1
openai-whisper==20250625
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
orjson==3.11.5
overrides==7.7.0
packaging==25.0
platformdirs==4.4.0
pooch==1.8.2
posthog==5.4.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pvporcupine==4.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pybase64==1.4.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
rapidfuzz>=3.0.0  # v32.3: Intelligent fuzzy text matching for OCR detection
redis>=4.2.0  # cost_tracker real-time streaming; redis.asyncio for async
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
prompt_toolkit>=3.0.43  # Non-blocking async REPL for Ouroboros CLI
# Optional TUI -- required for --ui tui
textual>=0.47.0  # v8.0.0 installed; TUI mode (--ui tui)
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.6.1
scipy>=1.13.1,<2
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soxr==1.0.0
speechbrain==1.0.3
starlette==0.49.3
sympy>=1.13.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers>=0.15.2,<1
tomli==2.3.0
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
transformers>=4.36.0,<5
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.3.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
webrtcvad==2.0.10
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0

# PostgreSQL drivers for CloudSQL
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
cloud-sql-python-connector>=1.7.0

# GCP Secret Manager
google-cloud-secret-manager>=2.20.0

# GCP Compute Engine (for Spot VM auto-creation)
google-cloud-compute>=1.40.0

# GCP Monitoring API (for real VM CPU metrics -- hypervisor-level, no agent needed)
google-cloud-monitoring>=2.0.0

# Python 3.9.x is EOL (2025-10). Python 3.11.x is EOL (2026-10). Minimum supported version is now 3.12.
# Upgrade your interpreter to Python 3.12 or 3.13 before installing.
# Requires-Python: >=3.12
